+++ pkg/resources/endpoints.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package utils
+package resources
 
 import (
 	corev1 "k8s.io/api/core/v1"
